pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django_security_questions"
version = "0.1.0"
description = "Security Questions for your django app"
authors = [{ name = "Piyush Khurana", email = "pka420@proton.me" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.6"
dependencies = ["django", "djangorestframework"]
classifiers = [
  "Development Status :: 3 - Alpha",

  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",

  "License :: OSI Approved :: MIT License",

  "Programming Language :: Python :: 3.6",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
]

